requirements.txt
alabaster==0.7.12
asgiref==3.5.2
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==1.0.0
Django==4.1
django-haystack==3.2.1
django-heroku==0.3.1
django-sphinxdoc==2.0.0
django-webpack-loader==1.6.0
docutils==0.19
gunicorn==20.1.0
idna==3.3
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.3
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
django-webpack-loader==1.6.0
pytest==7.1.2
pytest-django==4.5.2
django-allauth==0.51.0

PyJWT~=2.4.0
djangorestframework==3.13.1